refactor : structuring project
- 프로젝트를 구조화한다
- model, service, repository, exception 폴더를 추가한다
