+++ src/main/java/com/example/univeus/domain/chat/model/ChatMember.java
@@ -1,22 +1,25 @@
-package com.example.univeus.domain.chat;
+package com.example.univeus.domain.chat.model;
 
-import com.example.univeus.domain.meeting.Post;
+import com.example.univeus.domain.member.model.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatRoom {
+public class ChatMember {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @OneToOne
-    private Post post;
+    @ManyToOne
+    private Member member;
+
+    @ManyToOne
+    private ChatRoom chatRoom;
 }
