+++ src/main/java/com/example/univeus/domain/chat/model/ChatRoom.java
@@ -1,25 +1,22 @@
-package com.example.univeus.domain.meeting;
+package com.example.univeus.domain.chat.model;
 
+import com.example.univeus.domain.meeting.model.Post;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostImage {
+public class ChatRoom {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
+    @OneToOne
     private Post post;
-
-    @Lob
-    private String uri;
 }
