+++ src/main/java/com/example/univeus/domain/meeting/model/Location.java
@@ -1,5 +1,6 @@
-package com.example.univeus.domain.meeting;
+package com.example.univeus.domain.meeting.model;
 
+import com.example.univeus.domain.meeting.model.Coordinate;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.Embedded;
