+++ src/main/java/com/example/univeus/domain/meeting/model/PostImage.java
@@ -1,25 +1,26 @@
-package com.example.univeus.domain.chat;
+package com.example.univeus.domain.meeting.model;
 
-import com.example.univeus.domain.member.model.Member;
+import com.example.univeus.domain.meeting.model.Post;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatMessage {
+public class PostImage {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
-    private ChatRoom chatRoom;
+    private Post post;
 
-    @ManyToOne
-    private Member sender;
+    @Lob
+    private String uri;
 }
